[iOS][core] Migrate all remaining tests to Swift Testing
