+++ packages/expo-modules-core/ios/Tests/FunctionWithConvertiblesTests.swift
@@ -1,0 +1,156 @@
+// Copyright 2018-present 650 Industries. All rights reserved.
+
+import CoreGraphics
+import Testing
+
+@testable import ExpoModulesCore
+
+@Suite("FunctionWithConvertibles")
+struct FunctionWithConvertiblesTests {
+  let appContext: AppContext
+  let functionName = "function"
+
+  init() {
+    appContext = AppContext.create()
+  }
+
+  @Test
+  func `converts arguments to CoreGraphics types`() async {
+    let x = 18.3
+    let y = -4.1
+    let width = 734.6
+    let height = 592.1
+
+    await withCheckedContinuation { continuation in
+      mockModuleHolder(appContext) {
+        AsyncFunction(functionName) { (point: CGPoint, size: CGSize, vector: CGVector, rect: CGRect) in
+          #expect(Double(point.x) == x)
+          #expect(Double(point.y) == y)
+          #expect(Double(size.width) == width)
+          #expect(Double(size.height) == height)
+          #expect(Double(vector.dx) == x)
+          #expect(Double(vector.dy) == y)
+          #expect(Double(rect.origin.x) == x)
+          #expect(Double(rect.origin.y) == y)
+          #expect(Double(rect.width) == width)
+          #expect(Double(rect.height) == height)
+          continuation.resume()
+        }
+      }
+      .call(function: functionName, args: [
+        [x, y], // point
+        ["width": width, "height": height], // size
+        ["dx": x, "dy": y], // vector
+        [x, y, width, height] // rect
+      ])
+    }
+  }
+
+  @Test
+  func `converts arguments to CGColor`() async {
+    func testColorComponents(_ color: CGColor, _ red: CGFloat, _ green: CGFloat, _ blue: CGFloat, _ alpha: CGFloat) {
+      #expect(color.components?[0] == red / 255.0)
+      #expect(color.components?[1] == green / 255.0)
+      #expect(color.components?[2] == blue / 255.0)
+      #expect(color.components?[3] == alpha / 255.0)
+    }
+
+    await withCheckedContinuation { continuation in
+      mockModuleHolder(appContext) {
+        AsyncFunction(functionName) { (color1: CGColor, color2: CGColor, color3: CGColor, color4: CGColor) in
+          testColorComponents(color1, 0x2A, 0x4B, 0x5D, 0xFF)
+          testColorComponents(color2, 0x11, 0xFF, 0x00, 0xDD)
+          testColorComponents(color3, 0x66, 0x00, 0xCC, 0xAA)
+          testColorComponents(color4, 0x00, 0x00, 0x00, 0x00)
+          continuation.resume()
+        }
+      }
+      .call(function: functionName, args: [
+        "#2A4B5D",
+        0xDD11FF00,
+        "60CA",
+        0
+      ])
+    }
+  }
+
+  // MARK: - Record context
+
+  @Suite("Record")
+  struct RecordTests {
+    let appContext: AppContext
+    var runtime: ExpoRuntime {
+      get throws {
+        try appContext.runtime
+      }
+    }
+
+    init() {
+      appContext = AppContext.create()
+      appContext.moduleRegistry.register(holder: mockModuleHolder(appContext) {
+        Name("TestModule")
+
+        Function("withConvertibles") { TestRecord() }
+      })
+    }
+
+    @Test
+    func `converts convertibles`() throws {
+      let object = try runtime.eval("globalThis.result = expo.modules.TestModule.withConvertibles()")
+      #expect(object.kind == .object)
+
+      #expect(object.getObject().hasProperty("url") == true)
+      #expect(object.getObject().hasProperty("cgFloat") == true)
+      #expect(object.getObject().hasProperty("cgPoint") == true)
+      #expect(object.getObject().hasProperty("cgSize") == true)
+      #expect(object.getObject().hasProperty("cgVector") == true)
+      #expect(object.getObject().hasProperty("cgRect") == true)
+      #expect(object.getObject().hasProperty("cgColor") == true)
+
+      #expect(object.getObject().getProperty("url").kind == .string)
+      #expect(object.getObject().getProperty("url").getString() == "https://expo.dev/")
+
+      #expect(object.getObject().getProperty("cgFloat").kind == .number)
+      #expect(object.getObject().getProperty("cgFloat").getDouble() == 1.0)
+
+      let cgPoint = object.getObject().getProperty("cgPoint")
+      #expect(cgPoint.kind == .object)
+      #expect(cgPoint.getObject().getProperty("x").getDouble() == 1.0)
+      #expect(cgPoint.getObject().getProperty("y").getDouble() == 2.0)
+
+      let cgSize = object.getObject().getProperty("cgSize")
+      #expect(cgSize.kind == .object)
+      #expect(cgSize.getObject().getProperty("width").getDouble() == 100.0)
+      #expect(cgSize.getObject().getProperty("height").getDouble() == 200.0)
+
+      let cgVector = object.getObject().getProperty("cgVector")
+      #expect(cgVector.kind == .object)
+      #expect(cgVector.getObject().getProperty("dx").getDouble() == 100.0)
+      #expect(cgVector.getObject().getProperty("dy").getDouble() == 200.0)
+
+      let cgRect = object.getObject().getProperty("cgRect")
+      #expect(cgRect.kind == .object)
+      #expect(cgRect.getObject().getProperty("x").getDouble() == 50.0)
+      #expect(cgRect.getObject().getProperty("y").getDouble() == 100.0)
+      #expect(cgRect.getObject().getProperty("width").getDouble() == 200.0)
+      #expect(cgRect.getObject().getProperty("height").getDouble() == 300.0)
+
+      #expect(object.getObject().getProperty("cgColor").kind == .string)
+      #expect(object.getObject().getProperty("cgColor").getString() == "#ff0000ff")
+
+      #expect(object.getObject().getProperty("uiColor").kind == .string)
+      #expect(object.getObject().getProperty("uiColor").getString() == "#00ff00ff")
+    }
+  }
+}
+
+private struct TestRecord: Record {
+  @Field var url: URL = URL.init(string: "https://expo.dev/")!
+  @Field var cgFloat: CGFloat = CGFloat(1.0)
+  @Field var cgPoint: CGPoint = CGPoint(x: 1, y: 2)
+  @Field var cgSize: CGSize = CGSize(width: 100, height: 200)
+  @Field var cgVector: CGVector = CGVector(dx: 100, dy: 200)
+  @Field var cgRect: CGRect = CGRect(x: 50, y: 100, width: 200, height: 300)
+  @Field var cgColor: CGColor = UIColor.red.cgColor
+  @Field var uiColor: UIColor = UIColor.green
+}
